+++ src/utils/index.ts
@@ -1,0 +1,52 @@
+/*
+ * Copyright 2023 gematik GmbH
+ *
+ * The Authenticator App is licensed under the European Union Public Licence (EUPL); every use of the Authenticator App
+ * Sourcecode must be in compliance with the EUPL.
+ *
+ * You will find more details about the EUPL here: https://joinup.ec.europa.eu/collection/eupl
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the EUPL is distributed on an "AS
+ * IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the EUPL for the specific
+ * language governing permissions and limitations under the License.ee the Licence for the specific language governing
+ * permissions and limitations under the Licence.
+ */
+
+import cryptoJS from 'crypto-js'
+import base64url from 'base64url'
+
+/**
+ * creates random string
+ * @param size
+ */
+export function createRandomString(size = 50) {
+  const chars = '0123456789ABCDEFGHIJKLMNOPQRSTUVWXTZabcdefghiklmnopqrstuvwxyz'
+  let randomstring = ''
+  for (let i = 0; i < size; i++) {
+    const rNum = Math.floor(Math.random() * chars.length)
+    randomstring += chars.substring(rNum, rNum + 1)
+  }
+  return randomstring
+}
+
+export function createCodeChallenge(codeVerifier: string) {
+  const base64Digest = cryptoJS.SHA256(codeVerifier).toString(cryptoJS.enc.Base64)
+
+  return base64url.fromBase64(base64Digest)
+}
+
+export function parseJwt(token: string) {
+  const base64Url = token.split('.')[1]
+  const base64 = base64Url.replace(/-/g, '+').replace(/_/g, '/')
+  const jsonPayload = decodeURIComponent(
+    window
+      .atob(base64)
+      .split('')
+      .map(function (c) {
+        return '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2)
+      })
+      .join('')
+  )
+
+  return JSON.parse(jsonPayload)
+}