+++ ReleaseNotes.md
@@ -1,6 +1,9 @@
-<img align="right" width="200" height="37" src="src/assets/Gematik_Logo_Flag.png"/> <br/>
+<div style='text-align: right; margin-top: 30px'>
+  <img alt='Gematik Logo' style='height: 37px' src="src/assets/Gematik_Logo_Flag.png"/>
+</div>
 
 # Release Vue-Oidcauth-Sample
 
-## Release 4.0.0 (2023-03-22)
-- Redirect to webapplication as alternative to browseropen
+## Release 4.1.0
+* Combined Resource Server & Client
+* Moved from native Vue to Nuxt 3